src/main/webapp/app/entities/jhipsterSampleApplication/bom-source/route/bom-source-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IBomSource, BomSource } from '../bom-source.model';
import { BomSourceService } from '../service/bom-source.service';

import { BomSourceRoutingResolveService } from './bom-source-routing-resolve.service';

describe('BomSource routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: BomSourceRoutingResolveService;
  let service: BomSourceService;
  let resultBomSource: IBomSource | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(BomSourceRoutingResolveService);
    service = TestBed.inject(BomSourceService);
    resultBomSource = undefined;
  });

  describe('resolve', () => {
    it('should return IBomSource returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultBomSource = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultBomSource).toEqual({ id: 123 });
    });

    it('should return new IBomSource if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultBomSource = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultBomSource).toEqual(new BomSource());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as BomSource })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultBomSource = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultBomSource).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
